Move code out of ramp and into other modules
